[KIAI-32] ref: created separated types
