+++ src/global/api/api.consts.ts
