+++ src/global/store/index.ts
@@ -1,8 +1,7 @@
-import { authApi } from '@modules/Auth/shared/api/auth.api';
-
 import { LoginReducer as loginReducer } from './slices/login.slice';
 import { ModalReducer as modalReducer } from './slices/modal.slice';
 import { UserReducer as userReducer } from './slices/user.slice';
+import { authApi } from '@global/api/auth.api';
 import { configureStore } from '@reduxjs/toolkit';
 
 export const store = configureStore({
